Update tar.js
On windows and on android I am having trouble executing $npm install <tarball> . To resolve this issue, I set the entry.path = entry._header.path in the extractEntry function. Verison of npm on android is 2.15.1, Version of npm on windows is 3.10.8

Please verify the correct location of the fix, it should likely be in another location. But this does solve the problem on both systems.